Add files via upload
Technical and User manual templates
Please let me know if you require changes in the documents. The purpose is to fill in the information about your app development from Technical and User perspective. Once you add the information then we can work to formalize the format and style of the document. 
Thanks
Tahir